backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
django-redis>=5.2.0
pandas>=2.0.0
openpyxl>=3.1.0
